+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Added
 
+- [Enhancement] Add Import Capability for externally run Search Evaluation ([#46](https://github.com/opensearch-project/search-relevance/issues/46))
+
 ### Removed
 
 ### Fixed